Add tests, remove unnecessary functions, update return values in optimization functions

+++ src/zimscraperlib/image/optimization.py
@@ -30,25 +30,18 @@ import tempfile
 from typing import Optional, Tuple, Union
 
 import piexif
-from optimize_images.img_aux_processing import (do_reduce_colors,
-                                                make_grayscale,
-                                                rebuild_palette)
-from optimize_images.img_aux_processing import \
-    remove_transparency as remove_alpha
+from optimize_images.img_aux_processing import (
+    do_reduce_colors,
+    make_grayscale,
+    rebuild_palette,
+)
+from optimize_images.img_aux_processing import remove_transparency as remove_alpha
 from optimize_images.img_dynamic_quality import jpeg_dynamic_quality
 from PIL import Image
 
 from .convertion import convert_image
 from .probing import format_for
 from .utils import save_image
-
-
-def get_temporary_copy(src: pathlib.Path) -> pathlib.Path:
-    tmp_fh = tempfile.NamedTemporaryFile(delete=False, suffix=src.suffix)
-    tmp_fh.close()
-    tmp_path = pathlib.Path(tmp_fh.name)
-    shutil.copy(src, tmp_path)
-    return tmp_path
 
 
 def ensure_matches(
@@ -63,7 +56,7 @@ def ensure_matches(
 
 def optimize_png(
     src: Union[pathlib.Path, io.BytesIO],
-    dst: pathlib.Path,
+    dst: Optional[pathlib.Path] = None,
     reduce_colors: Optional[bool] = False,
     max_colors: Optional[int] = 256,
     fast_mode: Optional[bool] = True,
@@ -71,7 +64,7 @@ def optimize_png(
     background_color: Optional[Tuple[int, int, int]] = (255, 255, 255),
     grayscale: Optional[bool] = False,
     **options,
-) -> bool:
+) -> Tuple[bool, Union[pathlib.Path, io.BytesIO]]:
 
     """method to optimize PNG files using a pure python external optimizer
 
@@ -94,7 +87,7 @@ def optimize_png(
     img = Image.open(src)
     orig_mode = img.mode
 
-    if orig_mode == 'P':
+    if orig_mode == "P":
         final_colors = orig_colors = len(img.getcolors())
 
     result_format = "PNG"
@@ -110,22 +103,23 @@ def optimize_png(
     if not fast_mode and img.mode == "P":
         img, final_colors = rebuild_palette(img)
 
-    if isinstance(src, io.BytesIO) and dst is None:
-        img.save(dst, optimize=True, format=result_format)
-    else:
-        img.save(dst, optimize=True, format=result_format)
-    return True
+    if dst is None:
+        dst = io.BytesIO()
+    img.save(dst, optimize=True, format=result_format)
+    if isinstance(dst, io.BytesIO):
+        dst.seek(0)
+    return True, dst
 
 
 def optimize_jpeg(
     src: Union[pathlib.Path, io.BytesIO],
-    dst: pathlib.Path,
+    dst: Optional[pathlib.Path] = None,
     quality: Optional[int] = 85,
     fast_mode: Optional[bool] = True,
     keep_exif: Optional[bool] = True,
     grayscale: Optional[bool] = False,
     **options,
-) -> bool:
+) -> Tuple[bool, Union[pathlib.Path, io.BytesIO]]:
 
     """method to optimize JPEG files using a pure python external optimizer
     quality: JPEG quality (integer between 1 and 100)
@@ -140,11 +134,15 @@ def optimize_jpeg(
     ensure_matches(src, "JPEG")
 
     img = Image.open(src)
-    orig_size = os.path.getsize(src)
+    orig_size = (
+        os.path.getsize(src)
+        if isinstance(src, pathlib.Path)
+        else src.getbuffer().nbytes
+    )
 
     result_format = "JPEG"
     try:
-        had_exif = True if piexif.load(src)['Exif'] else False
+        had_exif = True if piexif.load(src)["Exif"] else False
     except piexif.InvalidImageDataError:  # Not a supported format
         had_exif = False
     except ValueError:
@@ -163,36 +161,34 @@ def optimize_jpeg(
     else:
         quality_setting, jpgdiff = jpeg_dynamic_quality(img)
 
-    if isinstance(src, io.BytesIO) and dst is None:
-        img.save(
-            src,
-            quality=quality_setting,
-            optimize=True,
-            progressive=use_progressive_jpg,
-            format=result_format)
-        src.seek(0)
-    else:
-        img.save(
-            dst,
-            quality=quality_setting,
-            optimize=True,
-            progressive=use_progressive_jpg,
-            format=result_format)
+    if dst is None:
+        dst = io.BytesIO()
+
+    img.save(
+        dst,
+        quality=quality_setting,
+        optimize=True,
+        progressive=use_progressive_jpg,
+        format=result_format,
+    )
+
+    if isinstance(dst, io.BytesIO):
+        dst.seek(0)
 
     if keep_exif and had_exif:
         piexif.transplant(os.path.expanduser(src), dst)
 
-    return True
+    return True, dst
 
 
 def optimize_webp(
-    src: pathlib.Path,
-    dst: pathlib.Path,
+    src: Union[pathlib.Path, io.BytesIO],
+    dst: Optional[pathlib.Path] = None,
     lossless: Optional[bool] = False,
     quality: Optional[int] = 60,
     method: Optional[int] = 6,
     **options,
-) -> bool:
+) -> Tuple[bool, Union[pathlib.Path, io.BytesIO]]:
     """method to optimize WebP using Pillow options
     lossless: Whether to use lossless compression (boolean)
         values: True | False
@@ -210,14 +206,19 @@ def optimize_webp(
         "method": method,
     }
 
-    try:
-        webp_image = Image.open(src)
-        save_image(webp_image, dst, fmt="WEBP", **params)
-    except Exception as exc:
-        if src.resolve() != dst.resolve() and dst.exists():
-            dst.unlink()
-        raise exc
-    return True
+    webp_image = Image.open(src)
+    if dst is None:
+        dst = io.BytesIO()
+        webp_image.save(dst, format="WEBP", **params)
+        dst.seek(0)
+    else:
+        try:
+            save_image(webp_image, dst, fmt="WEBP", **params)
+        except Exception as exc:
+            if src.resolve() != dst.resolve() and dst.exists():
+                dst.unlink()
+            raise exc
+    return True, dst
 
 
 def optimize_gif(
@@ -229,7 +230,7 @@ def optimize_gif(
     no_extensions: Optional[bool] = True,
     max_colors: Optional[int] = None,
     **options,
-) -> bool:
+) -> Tuple[bool, pathlib.Path]:
     """method to optimize GIFs using gifsicle >= 1.92
     optimize_level: Optimization level; higher values give better compression (integer between 1 and 3)
         values: 1 | 2 | 3
@@ -268,7 +269,7 @@ def optimize_gif(
 
     # raise error if unsuccessful
     gifsicle.check_returncode()
-    return True
+    return True, dst
 
 
 def optimize_image(
@@ -296,7 +297,7 @@ def optimize_image(
     else:
         src_img = pathlib.Path(src)
 
-    optimized = {
+    optimized, _ = {
         "JPEG": optimize_jpeg,
         "PNG": optimize_png,
         "GIF": optimize_gif,
